Added binding for "apm load"
Still doesn't quite work for linking symbols,
seems that PyCSH itself must also be imported with RTLD_GLOBAL... or use libcsh/

Also temporarily disabled "py run", until we decide on the optionality of the init func (i.e apm_init)
