+++ pycsh.pyi
@@ -10,6 +10,7 @@ These provide an object-oriented interface to libparam, but are largely meant fo
 from __future__ import annotations
 
 from _typeshed import Self
+from types import ModuleType as _ModuleType
 from typing import \
     Any as _Any, \
     Iterable as _Iterable, \
@@ -707,6 +708,22 @@ def sps(from: int, to: int, filename: str, node: int = None, window: int = None,
     :param ack_count: rdp ack for each (default = 2 packets)
 
     :raises ProgramDiffError: See class docstring.
+    """
+
+def apm_load(path: str = '~/.local/lib/csh/', filename: str = None, stop_on_error: bool = False) -> dict[str, _ModuleType | Exception]:
+    """
+    Loads both .py and .so APMs
+
+    Python built in 'import <apm>' machinery cannot load .so APMs that attempt to link with (Py)CSH.
+
+    :param path: Directory to scan for APMs to load
+    :param filename: Specific file/APM to load in the 'path' directory.
+    :param stop_on_error: Whether to stop and raise an exception when an APM fails to load.
+
+    :raises Exception: When stop_on_error==True and an APM fails to import.
+
+    :return: A dictionary with str paths of the modules as keys,
+        and the APM modules themselves as values (when loaded successfully, otherwise exception raised).
     """
 
 def csp_init(host: str = None, model: str = None, revision: str = None, version: int = 2, dedup: int = 3) -> None:
