+++ src/utils.h
@@ -12,19 +12,22 @@
 #define PY_SSIZE_T_CLEAN
 #include <Python.h>
 
+#include <dirent.h>
 #include <param/param.h>
 #include <param/param_queue.h>
 #include "parameter/pythonparameter.h"
 
 void cleanup_free(void *const* obj);
 void cleanup_str(char *const* obj);
+void _close_dir(DIR *const* dir);
 void cleanup_GIL(PyGILState_STATE * gstate);
 void cleanup_pyobject(PyObject **obj);
 
 void state_release_GIL(PyThreadState ** state);
 
-#define CLEANUP_STR __attribute__((cleanup(cleanup_str)))
 #define CLEANUP_FREE __attribute__((cleanup(cleanup_free)))
+#define CLEANUP_STR __attribute__((cleanup(cleanup_str)))
+#define CLEANUP_DIR __attribute__((cleanup(_close_dir)))
 #define CLEANUP_GIL __attribute__((cleanup(cleanup_GIL)))
 #define AUTO_DECREF __attribute__((cleanup(cleanup_pyobject)))
